"Download video" using Celery worker - feature added.

+++ backend/downloads/models.py
@@ -10,28 +10,44 @@ class Download(models.Model):
     """
 
     class Status(models.TextChoices):
+        """
+        Status choices for Downloads.
+        """
+
         PENDING = "PG", _("Pending")
         COMPLETED = "CD", _("Completed")
         FAILED = "FD", _("Failed")
 
-    title = models.CharField(verbose_name=_("title"), max_length=512)
+    title = models.CharField(
+        verbose_name=_("title"),
+        max_length=512,
+        blank=True,
+        null=True,
+    )
     status = models.CharField(
         verbose_name=_("status"),
         max_length=2,
         choices=Status.choices,
         default=Status.PENDING,
     )
-    bookmark = models.ForeignKey(
+    bookmark = models.OneToOneField(
         verbose_name=_("bookmark"),
         to=Bookmark,
         on_delete=models.CASCADE,
-        related_name="downloads",
+        related_name="download",
         blank=True,
         null=True,
         default=None,
     )
-    file = models.FileField(verbose_name=_("file"))
-    file_size = models.IntegerField(verbose_name=_("file size, bytes"))
+    file = models.FileField(
+        verbose_name=_("file"),
+        null=True,
+        blank=True,
+    )
+    file_size = models.IntegerField(
+        verbose_name=_("file size, bytes"),
+        default=0,
+    )
     created = models.DateTimeField(verbose_name=_("created"), auto_now_add=True)
     updated = models.DateTimeField(verbose_name=_("updated"), auto_now=True)
 
@@ -41,11 +57,13 @@ class Download(models.Model):
         ordering = ["-created"]
 
     def __str__(self):
-        return self.title
+        return self.title if self.title else _("Untitled")
 
     def delete(self, *args, **kwargs):
         """
-        Delete file with Download instance.
+        Delete attached file before Download instance gets deleted.
+        See also `signals.py`.
         """
-        self.file.delete(save=False)
+        if self.file:
+            self.file.delete(save=False)
         super().delete(*args, **kwargs)
